+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 # Next Release
 
 - [#477](https://github.com/IAMconsortium/pyam/pull/477) Add a nightly test suite
+- [#471](https://github.com/IAMconsortium/pyam/pull/471) Add a `iiasa.Connection.properties()` function to retrieve scenario audit data
 
 # Release v0.10.0
 
@@ -27,6 +28,8 @@ are ordered based on the `run_control()['order']` dictionary by default.
 
 ## Individual updates
 
+- [#464](https://github.com/IAMconsortium/pyam/pull/463) Add `order` arg to `stackplot()`
+- [#466](https://github.com/IAMconsortium/pyam/pull/466) Add a `sankey()`function to create sankey diagrams
 - [#475](https://github.com/IAMconsortium/pyam/pull/475) Add a scatter plot example to the gallery
 - [#473](https://github.com/IAMconsortium/pyam/pull/473) Refactor to plotting API following pandas/matplotlib implementation
 - [#472](https://github.com/IAMconsortium/pyam/pull/472) Add a `sankey()` example to the plotting gallery